Update Unique Number of Occurrences
Question :
Leetcode 1207. Unique Number of Occurrences 
Given an array of integers arr, return true if the number of occurrences of each value in the array is unique, or false otherwise.
*Example : 
Input: arr = [1,2,2,1,1,3]
Output: true
Explanation: The value 1 has 3 occurrences, 2 has 2 and 3 has 1. No two values have the same number of occurrences.


Solution:
first store the number of occurrences of each element in the map.
as map has unique key value.

secondly take one set...traverse through second value of map..store it in the set if it does not exist before in set..if it exists before in set already then return false (which means occurrences is repeated.) 

count function of stl return the no of occurrence of element in set.